Add files via upload / First version of project
This is my first major project. I only learned about Github now sadly, it would have save me a lot of hassle. The goal is to create a website, which scrolls infinitely. In that website, iFrames should be integrated randomly, with random sizes. So far, I have done all of that. Now what needs to be changed is getting Cross Domain Access.